refactor: decompose window component
