+++ src/app/(main)/components/AppNavigator/index.styles.scss
@@ -11,6 +11,7 @@
   border-radius: 10px;
   border: 1px solid rgba(255, 255, 255, 0.18);
   padding: var(--space-md) var(--space-xl);
+  padding-bottom: var(--space-lg);
   border-radius: 16px;
   z-index: 100;
 
@@ -47,13 +48,12 @@
 
     .minimized-indicator {
       position: absolute;
-      top: -4px;
-      right: -4px;
-      width: 8px;
-      height: 8px;
-      background-color: #00ca4e;
+      bottom: -8px;
+      width: 5px;
+      height: 5px;
+      background-color: var(--grey-200);
       border-radius: 50%;
-      box-shadow: 0 0 4px rgba(0, 202, 78, 0.5);
+      box-shadow: 0 0 4px rgba(var(--grey-500), 0.5);
     }
   }
 }
